the position of the main script has been changed, message handlers are separated into different files, scripts renamed
