+++ src/config/dbconfig.py
@@ -6,8 +6,8 @@ USER_DATA_TABLES = ("users", "states", "users_info", "photos", "users_searching_
 
 OTHER_TABLES = ("main_messages", )
 
-DEFAULT_LOGGING_TYPE = TYPE_RECORDING[0]
+DEFAULT_RECORD_TYPE = TYPE_RECORDING[0]
 
-DEFAULT_LOGGING_STAGE = STAGES_RECORDING[0]
+DEFAULT_RECORD_STAGE = STAGES_RECORDING[0]
 
 AMOUNT_CITIES = 1138
