+++ src/prettybot/bot/dbassistant/registrations/registration_data_handler.py
@@ -1,9 +1,9 @@
 from typing import Union
 
-from src.prettybot.bot.db import database_assistant
+from src.prettybot.bot.dbassistant import database_assistant
 from src.prettybot.bot.minorscripts import supportive
-from src.prettybot.bot.db.registration import age_range_creator
-from src.prettybot.bot.db.registration import value_validator
+from src.prettybot.bot.dbassistant.registrations import age_range_creator
+from src.prettybot.bot.dbassistant.registrations import value_converter
 from src.config.recording_stages import *
 
 
@@ -14,6 +14,7 @@ class RegistrationParamsHandler:
         self._database_operation_assistant = database_operation_assistant
 
     def _save_user_photo(self, user_id: int, file_id: str, record_type: str) -> None:
+        print(record_type, '----')
         if record_type == TYPE_RECORDING[0]:
             self._database_operation_assistant.save_photo_id(user_id=int(user_id), file_id=str(file_id))
 
@@ -25,12 +26,11 @@ class RegistrationParamsHandler:
             user_id: int,
             record_type: int,
             record_stage: int) -> None:
-
+        print(record_stage, '---------------')
         if record_stage == PHOTO_STAGE:
             return self._save_user_photo(user_id=user_id, file_id=message_payload, record_type=record_type)
 
-        column_value = value_validator.validate_user_value(user_value=message_payload, record_strage=record_stage)
-
+        column_value = value_converter.convert_user_param_value(user_value=message_payload, record_strage=record_stage)
         if record_type != TYPE_RECORDING[2]:
             self._database_operation_assistant.record_user_param(user_id=user_id,
                                                                  name_param=NAME_COLUMN_BY_LOGSTAGE[record_stage],
