+++ src/prettybot/bot/dbassistant/sqltemplates.py
