+++ src/prettybot/bot/messages/handlers/user_answer_handlers.py
@@ -10,16 +10,34 @@ from src.config.pbconfig import *
 __all__ = ['handle_message']
 
 
-def handle_age(*args: tuple[str],
-               message_text: str,
-               user_lang_code: str) -> Union[str, dataclass.ResultOperation]:
-    recordtype = args[2]
-    if recordtype in TYPE_RECORDING[:2]:
+def handle_message(user_lang_code: str, record_stage: str, **kwargs) -> Union[dataclass.ResultOperation, str]:
+    """
+    kwargs values ---
+    1. message text / file id {file id in handle photo}
+    3. record_type / cities {handle age, handle city}
+    4. record_stage
+
+    :param record_stage: stage of recording
+    :param user_lang_code: user language for statement
+
+    :returns: your statement
+    """
+    try:
+        text_handler = HANDLE_SCRIPT_BY_RECORDING_STAGE[record_stage]
+        return text_handler(user_lang_code=user_lang_code, **kwargs)
+
+    except KeyError:
+        return
+
+
+def _handle_age(user_lang_code: str, **kwargs) -> Union[str, dataclass.ResultOperation]:
+    record_type = kwargs.get('record_type')
+    message_text = kwargs.get('message_text')
+
+    if record_type in TYPE_RECORDING[:2]:
         try:
-            age = int(message_text)
-            if not LOWER_AGE_LIMIT <= age <= UPPER_AGE_LIMIT:
-                return dataclass.ResultOperation(status=False, object=STATEMENTS_BY_LANG[
-                    user_lang_code].invalid_t_age)
+            if not LOWER_AGE_LIMIT <= int(message_text) <= UPPER_AGE_LIMIT:
+                return dataclass.ResultOperation(status=False, object=STATEMENTS_BY_LANG[user_lang_code].invalid_t_age)
 
             return dataclass.ResultOperation(status=True, object=STATEMENTS_BY_LANG[user_lang_code].q_city)
 
@@ -27,7 +45,7 @@ def handle_age(*args: tuple[str],
             return dataclass.ResultOperation(status=False,
                                              object=STATEMENTS_BY_LANG[user_lang_code].invalid_v_age)
 
-    elif recordtype == TYPE_RECORDING[2]:
+    elif record_type == TYPE_RECORDING[2]:
         if supportive.check_numbers_in_string(string=message_text) is not None:
             numbers = [i for i in message_text if i.isdigit()]
 
@@ -42,14 +60,12 @@ def handle_age(*args: tuple[str],
                                              object=STATEMENTS_BY_LANG[user_lang_code].invalid_t_range_age)
 
 
-def handle_city(*args, message_text: str, user_lang_code: str):
-    cities = [item for item in args if type(item) is dict and len(item) == AMOUNT_CITIES][0]
+def _handle_city(user_lang_code: str, **kwargs):
+    cities = kwargs.get('cities')
+    message_text = kwargs.get('message_text')
 
     if str.lower(str(message_text)) not in cities:
-
-        simular_cities = supportive.get_similar_cities(city=str.lower(message_text),
-                                                       cities=cities)
-
+        simular_cities = supportive.get_similar_cities(city=str.lower(message_text), cities=cities)
         if simular_cities:
             return dataclass.ResultOperation(status=False,
                                              object=supportive.generate_drop_down_cities_list(
@@ -61,11 +77,19 @@ def handle_city(*args, message_text: str, user_lang_code: str):
     return dataclass.ResultOperation(status=True, object=STATEMENTS_BY_LANG[user_lang_code].q_sex)
 
 
-def handle_sex(*args, **kwargs):
-    return dataclass.ResultOperation(status=False)
+def _handle_sex(user_lang_code: str, **kwargs):
+    record_type = kwargs.get('record_type')
+
+    if record_type == TYPE_RECORDING[0]:
+        return dataclass.ResultOperation(status=False, object=STATEMENTS_BY_LANG[user_lang_code].q_sex)
+
+    elif record_type == TYPE_RECORDING[2]:
+        return dataclass.ResultOperation(status=False, object=STATEMENTS_BY_LANG[user_lang_code].q_find_sex)
 
 
-def handle_description(*_, message_text: str, user_lang_code: str):
+def _handle_description(user_lang_code: str, **kwargs):
+    message_text = kwargs.get('message_text')
+
     if len(message_text) > MAX_LEN_DESCRIPTION:
         return dataclass.ResultOperation(status=False,
                                          object=STATEMENTS_BY_LANG[user_lang_code].invalid_l_desc.format(
@@ -75,32 +99,15 @@ def handle_description(*_, message_text: str, user_lang_code: str):
     return dataclass.ResultOperation(status=True, object=STATEMENTS_BY_LANG[user_lang_code].q_photo)
 
 
-def handle_photo(*_, **kwargs):
+def _handle_photo(user_lang_code: str, **_):
     return dataclass.ResultOperation(status=False,
-                                     object=STATEMENTS_BY_LANG[kwargs.get('user_lang_code')].invalid_t_photo)
-
-
-def handle_message(
-        *args,
-        recordstage: str) -> dataclass.ResultOperation:
-
-    """
-    args values ---
-    1. message text / file id {file id in handle photo}
-    2. user lang code
-    3. recordtype / cities {handle age, handle city}
-
-    :param args:
-    :param recordstage:
-    :return: your statement
-    """
-    return HANDLE_SCRIPT_BY_RECORDING_STAGE[recordstage](*args, message_text=args[0], user_lang_code=args[1])
+                                     object=STATEMENTS_BY_LANG[user_lang_code].invalid_t_photo)
 
 
 HANDLE_SCRIPT_BY_RECORDING_STAGE = {
-    'age': handle_age,
-    'cty': handle_city,
-    'sex': handle_sex,
-    'dsc': handle_description,
-    'pht': handle_photo,
+    'age': _handle_age,
+    'cty': _handle_city,
+    'sex': _handle_sex,
+    'dsc': _handle_description,
+    'pht': _handle_photo,
 }
